Deprecating bazel, jdk7, jdk8, jre7, jre8, p4broker, libsasl2
Bazel: This plan no longer builds, and hasn't for over a year fixing this would be time-cost prohibitive, combined with the fact that there are no rdeps for this plan, means that it meets the criteria for deprecation until such time as it is required to be a core plan again.
P4broker: This is a closed source application and shouldn't be incuded in core-plans.
JRE7/8 & JKD7/8: Licensing from oracle has changed and these will no longer be able to be redistributed. As an alternative, please check corretto in the core-plans.
Libsasl2: Removing libsasli2 because it is replaced by cyrus-sasl

gocd-server : Changing jre8 to corretto8

Signed-off-by: Steven Marshall <SMarshall@chef.io>